# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ubhi Backend API
  version: 1.0.0
  description: >
    API for the Ubhi brand site. Powers the public storefront (workshops, shop,
    gallery, journal, bookings, orders, newsletter and snail-mail signups) and
    the back-office admin dashboard (auth, users, customers, settings).
    Stateless JWT auth; role + per-section permission model. Storage is
    pluggable (zero-config JSON file store by default, PostgreSQL when
    DATABASE_URL is set).
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8090
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    AdminUser:
      type: object
      description: password_hash is never returned.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            type: string
        name:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            type: string
        name:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            type: string
        name:
          type: string
        active:
          type: boolean
    PasswordRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
...
    Workshop:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        blurb:
          type: string
        session_date:
          type: string
        price:
          type: number
        spots:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    WorkshopInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        blurb:
          type: string
        session_date:
          type: string
        price:
          type: number
        spots:
          type: integer
        image:
          type: string
        published:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        blurb:
          type: string
        price:
          type: number
        stock:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        blurb:
          type: string
        price:
          type: number
        stock:
          type: integer
        image:
          type: string
        published:
          type: boolean
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role or permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List admin users (owner only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: Create an admin user (owner only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    patch:
      tags: [Users]
      summary: Update an admin user (owner only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete an admin user (owner only; cannot delete last owner or self)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/password:
    post:
      tags: [Users]
      summary: Set a user's password (owner only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/workshops:
    get:
      tags: [Workshops]
      summary: List published workshops (public)
      responses:
        '200':
          description: Published workshops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workshop'
    post:
      tags: [Workshops]
      summary: Create a workshop (section 'workshops')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopInput'
      responses:
        '201':
          description: Created workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workshop'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/workshops/all:
    get:
      tags: [Workshops]
      summary: List all workshops incl. unpublished (section 'workshops')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All workshops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workshop'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/workshops/{id}:
    patch:
      tags: [Workshops]
      summary: Update a workshop (section 'workshops')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopInput'
      responses:
        '200':
          description: Updated workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workshop'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workshops]
      summary: Delete a workshop (section 'workshops')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/shop:
    get:
      tags: [Shop]
      summary: List published products (public)
      responses:
        '200':
          description: Published products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Shop]
      summary: Create a product (section 'shop')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/shop/all:
    get:
      tags: [Shop]
      summary: List all products incl. unpublished (section 'shop')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/shop/{id}:
    patch:
      tags: [Shop]
      summary: Update a product (section 'shop')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Shop]
      summary: Delete a product (section 'shop')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...